Fixed compilator warning

+++ absp/src/ADQ214.cpp
@@ -221,7 +221,6 @@ int ABCD::ADQ214::Configure()
     // unnecessary.
     CHECKZERO(ADQ_SetTestPatternMode(adq_cu_ptr, adq_num, 0));
 
-
     // -------------------------------------------------------------------------
     //  Channels settings
     // -------------------------------------------------------------------------
@@ -603,7 +602,7 @@ int ABCD::ADQ214::GetWaveformsFromCard(std::vector<struct event_waveform> &wavef
         std::cout << "Timestamp overflows: " << timestamp_overflows << "; ";
         std::cout << std::endl;
     }
-    
+
     //ADQ_GetOverflow(adq_cu_ptr, adq_num);
 
     return DIGITIZER_SUCCESS;
@@ -740,9 +739,9 @@ int ABCD::ADQ214::ReadConfig(json_t *config)
         std::cout << std::endl;
     }
 
-    ////////////////////////////////////////////////////////////////////////////
-    // Reading the trigger configuration                                      //
-    ////////////////////////////////////////////////////////////////////////////
+    // -------------------------------------------------------------------------
+    //  Reading the trigger configuration
+    // -------------------------------------------------------------------------
     json_t *trigger_config = json_object_get(config, "trigger");
 
     const char *cstr_trigger_source = json_string_value(json_object_get(trigger_config, "source"));
@@ -831,9 +830,9 @@ int ABCD::ADQ214::ReadConfig(json_t *config)
         std::cout << std::endl;
     }
 
-    ////////////////////////////////////////////////////////////////////////////
-    // Reading the single channels configuration                              //
-    ////////////////////////////////////////////////////////////////////////////
+    // -------------------------------------------------------------------------
+    //  Reading the single channels configuration
+    // -------------------------------------------------------------------------
     // First resetting the channels statuses
     for (unsigned int channel = 0; channel < GetChannelsNumber(); channel++) {
         SetChannelEnabled(channel, false);
@@ -852,7 +851,7 @@ int ABCD::ADQ214::ReadConfig(json_t *config)
             json_t *json_id = json_object_get(value, "id");
 
             if (json_id != NULL && json_is_integer(json_id)) {
-                const unsigned int id = json_integer_value(json_id);
+                const int id = json_integer_value(json_id);
 
                 if (GetVerbosity() > 0)
                 {
@@ -920,7 +919,7 @@ int ABCD::ADQ214::ReadConfig(json_t *config)
 
                 json_object_set_nocheck(value, "coupling", json_string(ADQ_descriptions::analog_front_end_coupling.at(analog_front_end_coupling).c_str()));
 
-                if (0 <= id && id < GetChannelsNumber()) {
+                if (0 <= id && id < static_cast<int>(GetChannelsNumber())) {
                     SetChannelEnabled(id, enabled);
                     SetChannelTriggering(id, triggering);
                     analog_front_end_couplings[id] = analog_front_end_coupling;
@@ -935,9 +934,9 @@ int ABCD::ADQ214::ReadConfig(json_t *config)
         }
     }
 
-    ////////////////////////////////////////////////////////////////////////////
-    // Reading the waveforms settings                                         //
-    ////////////////////////////////////////////////////////////////////////////
+    // -------------------------------------------------------------------------
+    //  Reading the waveforms settings
+    // -------------------------------------------------------------------------
 
     pretrigger = json_integer_value(json_object_get(config, "pretrigger"));
